Change page load file from server to global
